refactor(paths): writing a file is not a path concern
